pyproject.toml
[project]
name = "svhn-kmeans"
version = "0.1.0"
description = "End-to-end SVHN digit recognition system using K-means features and beam search"
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.7.0",
    "Pillow>=10.0.0",
    "h5py>=3.0.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
cnn = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
gpu-cuda12 = [
    "cupy-cuda12x>=12.0.0", 
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
